// apps/web/src/app/api/model/integrationCreateRequest.ts
/**
 * Talos API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface IntegrationCreateRequest {
    type: string;
    name: string;
    configJson?: { [key: string]: any; };
    /**
     * Plaintext credential (e.g. a GitHub fine-grained PAT); encrypted at rest and never echoed back.
     */
    secret?: string;
    /**
     * \'pat\', \'deploy_key\', ... (integration_credentials.auth_mode, Section 9.2)
     */
    authMode?: string;
}